pyproject.toml
[project]
name = "MLOps"
version = "20240504"
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26.4",
  "pandas>=2.2.2",
  "ruff",
  "pre-commit",
  "sktime",
  "u8darts"
]

[tool.ruff]
# Same as Black.
# Set the maximum line length to 88.
line-length = 88
indent-width = 4

[tool.ruff.lint]
# Enable flake8-bugbear (`B`) rules, Pyflakes (`F`) and a subset of the pycodestyle (`E`) codes by default.
select = ["E4", "E7", "E9", "F", "B"] 

# Avoid enforcing line-length violations (`E501`)
ignore = ["E501"]

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. 
docstring-code-format = false

# Set the line length limit used in docstrings.
docstring-code-line-length = "dynamic"

[tool.ruff.lint.flake8-quotes] 
docstring-quotes = "double"
